% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actflow_predict.R
\name{actflow_predict}
\alias{actflow_predict}
\title{Activity flow prediction}
\usage{
actflow_predict(
  act,
  fc,
  separate_by_target = FALSE,
  transfer = c("linear", "relu", "sigmoid", "logit")
)
}
\arguments{
\item{act}{Activation vector/matrix.}

\item{fc}{Connectivity matrix (nodes x nodes).}

\item{separate_by_target}{If TRUE, act is interpreted as (target x source).}

\item{transfer}{Transfer function name.}
}
\description{
Activity flow prediction
}
