% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actflow_test.R
\name{actflow_test}
\alias{actflow_test}
\title{Test actflow predictions across conditions/subjects}
\usage{
actflow_test(
  act_group,
  fc_group,
  act_group_test = NULL,
  comparison = "fullcompare_compthenavg",
  transfer = c("linear", "relu", "sigmoid", "logit"),
  use_cpp = TRUE
)
}
\arguments{
\item{act_group}{Array with dimensions (nodes x conditions x subjects).}

\item{fc_group}{Array with dimensions (nodes x nodes x subjects) or
(nodes x nodes x conditions x subjects).}

\item{act_group_test}{Optional test activations.}

\item{comparison}{Comparison type.}

\item{transfer}{Transfer function for actflow prediction.}

\item{use_cpp}{Use native kernels when available.}
}
\description{
Test actflow predictions across conditions/subjects
}
