% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fc.R
\name{estimate_fc}
\alias{estimate_fc}
\title{Estimate connectivity matrix}
\usage{
estimate_fc(
  x,
  method = c("corr", "multreg", "partial", "combined", "lasso_cv", "glasso_cv"),
  partial_method = c("ridge", "glasso", "pc"),
  ...
)
}
\arguments{
\item{x}{Numeric matrix.}

\item{method}{FC method.}

\item{partial_method}{Partial estimator when \code{method = "partial"}.}

\item{...}{Method-specific args.}
}
\description{
Estimate connectivity matrix
}
