% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_corr.R
\name{estimate_fc_corr}
\alias{estimate_fc_corr}
\title{Correlation connectivity}
\usage{
estimate_fc_corr(
  x,
  orientation = c("nodes_by_time", "time_by_nodes"),
  use_cpp = TRUE
)
}
\arguments{
\item{x}{Numeric matrix or 3D array.}

\item{orientation}{Orientation of matrix.}

\item{use_cpp}{Use native kernel when available.}
}
\value{
FC matrix (nodes x nodes) or 3D FC array (nodes x nodes x subjects).
}
\description{
Correlation connectivity
}
