% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_glasso.R
\name{estimate_fc_glasso}
\alias{estimate_fc_glasso}
\title{Compute graphical-lasso partial-correlation FC}
\usage{
estimate_fc_glasso(
  x,
  lambda,
  orientation = c("nodes_by_time", "time_by_nodes"),
  zscore = TRUE
)
}
\arguments{
\item{x}{Numeric matrix.}

\item{lambda}{Positive scalar lambda.}

\item{orientation}{Matrix orientation.}

\item{zscore}{Whether to z-score node time series before estimation.}
}
\value{
Partial-correlation FC matrix.
}
\description{
Compute graphical-lasso partial-correlation FC
}
