% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_lasso.R
\name{estimate_fc_lasso}
\alias{estimate_fc_lasso}
\title{Compute lasso FC with fixed lambda}
\usage{
estimate_fc_lasso(
  x,
  lambda,
  orientation = c("nodes_by_time", "time_by_nodes"),
  zscore = TRUE
)
}
\arguments{
\item{x}{Numeric matrix.}

\item{lambda}{Lambda scalar or vector.}

\item{orientation}{Matrix orientation.}

\item{zscore}{Whether to z-score node time series before fitting.}
}
\value{
FC matrix or FC array with dimensions (nodes x nodes x n_lambda).
}
\description{
Compute lasso FC with fixed lambda
}
