% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_multreg.R
\name{estimate_fc_multreg}
\alias{estimate_fc_multreg}
\title{Multiple-regression connectivity}
\usage{
estimate_fc_multreg(
  x,
  orientation = c("nodes_by_time", "time_by_nodes"),
  ridge = 0,
  use_cpp = TRUE
)
}
\arguments{
\item{x}{Numeric matrix.}

\item{orientation}{Orientation of matrix.}

\item{ridge}{Optional ridge loading added to covariance diagonal.}

\item{use_cpp}{Use native kernel when available.}
}
\value{
FC matrix (nodes x nodes) with rows as targets and columns as sources.
}
\description{
Multiple-regression connectivity
}
