% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_model.R
\name{model_compare}
\alias{model_compare}
\title{Compare model predictions to target activations}
\usage{
model_compare(
  target,
  model1,
  model2 = NULL,
  comparison = c("fullcompare_compthenavg", "conditionwise_compthenavg",
    "conditionwise_avgthencomp", "nodewise_compthenavg", "nodewise_avgthencomp"),
  use_cpp = TRUE
)
}
\arguments{
\item{target}{Array with dimensions (nodes x conditions x subjects).}

\item{model1}{Array with dimensions (nodes x conditions x subjects).}

\item{model2}{Optional model2 array.}

\item{comparison}{One of \code{fullcompare_compthenavg},
\code{conditionwise_compthenavg}, \code{conditionwise_avgthencomp},
\code{nodewise_compthenavg}, or \code{nodewise_avgthencomp}.}

\item{use_cpp}{Use native kernels when available.}
}
\description{
Compare model predictions to target activations
}
