% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_ceiling.R
\name{noise_ceiling}
\alias{noise_ceiling}
\title{Noise ceiling estimate}
\usage{
noise_ceiling(act_run1, act_run2, ...)
}
\arguments{
\item{act_run1}{Array with dimensions (nodes x conditions x subjects).}

\item{act_run2}{Array with dimensions (nodes x conditions x subjects).}

\item{...}{Passed to model_compare.}
}
\description{
Noise ceiling estimate
}
