% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_hdf5.R
\name{read_actflower_h5}
\alias{read_actflower_h5}
\title{Read an HDF5 dataset}
\usage{
read_actflower_h5(path, dataset)
}
\arguments{
\item{path}{File path.}

\item{dataset}{Dataset name/path in the HDF5 file.}
}
\value{
Numeric array/matrix.
}
\description{
Read an HDF5 dataset
}
