% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_hdf5.R
\name{write_actflower_h5}
\alias{write_actflower_h5}
\title{Write an HDF5 dataset}
\usage{
write_actflower_h5(path, name, x)
}
\arguments{
\item{path}{File path.}

\item{name}{Dataset name/path.}

\item{x}{Object to write.}
}
\description{
Write an HDF5 dataset
}
