% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_corr.R
\name{estimate_fc_corr}
\alias{estimate_fc_corr}
\title{Correlation connectivity}
\usage{
estimate_fc_corr(x, orientation = c("nodes_by_time", "time_by_nodes"))
}
\arguments{
\item{x}{Numeric matrix.}

\item{orientation}{Orientation of matrix.}
}
\value{
FC matrix (nodes x nodes).
}
\description{
Correlation connectivity
}
